Rename jsx-sort-prop-types to sort-prop-types
This rule has nothing to do with JSX, so it doesn't make sense for "jsx"
to appear in its name.

To prevent breaking this rule for people, I have set up
jsx-sort-prop-types as an alias with a deprecation warning. Eventually,
we should remove it altogether. For the deprecation warning, I borrowed
code from jsx-quotes.

While I was at it, I corrected some minor wording issues in the
documentation for this rule (mostly using "propTypes" instead of
"propsTypes").

Fixes #87.
