+++ index.js
@@ -25,6 +25,7 @@ module.exports = {
     'jsx-curly-spacing': require('./lib/rules/jsx-curly-spacing'),
     'jsx-equals-spacing': require('./lib/rules/jsx-equals-spacing'),
     'jsx-sort-props': require('./lib/rules/jsx-sort-props'),
+    'sort-prop-types': require('./lib/rules/sort-prop-types'),
     'jsx-sort-prop-types': require('./lib/rules/jsx-sort-prop-types'),
     'jsx-boolean-value': require('./lib/rules/jsx-boolean-value'),
     'sort-comp': require('./lib/rules/sort-comp'),
@@ -66,6 +67,7 @@ module.exports = {
     'jsx-curly-spacing': 0,
     'jsx-equals-spacing': 0,
     'jsx-sort-props': 0,
+    'sort-prop-types': 0,
     'jsx-sort-prop-types': 0,
     'jsx-boolean-value': 0,
     'sort-comp': 0,
